requirements.txt
django
python-binance